Add tests for POST /auth/login endpoint
